refactor(internal/rust): merge rust packages
Merge internal/librarian/rust into internal/librarian/internal/rust to
consolidate all Rust-related functionality into a single package.

Remove typo checks from internal/sidekick to maintain feature parity.

Consolidate Create helper functions into a single implementation to
reduce unnecessary indirection.
