--- internal/librarian/rust/testdata/new-lib-format/src/main.rs
@@ -1,4 +1,0 @@
-fn main(){
-let x=5;
-    if x==5{println!("bad spacing");}
-}
